Create Transmits data as a string
This is the code for the Teensy to send the data from the sensors as a string.  It can be modified easily to also print it to the serial monitor.  The GPS code needs to be cleaned up for it to compile, so that part has been commented out, along with the SD card code.